requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0

# LLM and NLP
transformers>=4.20.0
sentence-transformers>=2.2.0
openai>=0.27.0
langchain>=0.0.200
langchain-openai>=0.0.1
langchain-community>=0.0.1

# Vector Databases
faiss-cpu>=1.7.0
chromadb>=0.3.0
pinecone-client>=2.2.0
weaviate-client>=3.0.0

# Web and API
fastapi>=0.95.0
uvicorn>=0.20.0
streamlit>=1.20.0
gradio>=3.20.0

# MLOps
mlflow>=2.0.0
wandb>=0.15.0
ray>=2.0.0

# Utilities
python-dotenv>=0.20.0
tqdm>=4.64.0
pytest>=7.0.0
